Client - minor fixes
* Discounts: Fixed web API server did not accept double for the min and max values
* Discounts: Fixed client was not displaying the min and max values appropriately to the condition type of the discount
* Policies: Added a key for each new policy so it can serve as a key to be given to React.

+++ TradingSystem/TradingSystem.WebApi/DTO/Store/Discounts/DiscountParamsDTO.cs
@@ -21,8 +21,8 @@ namespace TradingSystem.WebApi.DTO.Store.Discounts
 
         public string? ConditionType { get; set; }
 
-        public int? MinValue { get; set; }
-        public int? MaxValue { get; set; }
+        public double? MinValue { get; set; }
+        public double? MaxValue { get; set; }
 
         public DateTime? StartDate { get; set; }
         public DateTime? EndDate { get; set; }
